+ Adds a Task VO to represent a task. ~ Fixes some tests. No reason to add a test_task just yet.
~ I've also decided to write long commit messages in order to do an experiment:
I'll add as much context as possible to the commit messages so that others can follow my decisions.
I also think it will be pretty cool to read these when I am git-bisecting for regressions :)

FYI: these "sections" be will at the bottom of each commit, thus it should not interfere in the speed
of reading.