+ Task model integration with sqlalchemy ~ TaskManagement controller contains some logic to verify how sqlalchemy would deal with the insertion. ~ Base model with silly db connection.
~
Feels good to be back to this project after a very busy beginning of the week. Its been 4 years since my last sqlAlchemy adventure. I've been playing with my models a struggling to make the most simple things work. At least I got a dummy controller that persist information across sessions for the lifespan of the app.

This commit has no tests since it is more of a free-for-all session to have a feeling on how things should go around. Will drive my refactor through tests in the upcoming coding sessions.

PS: as expected, barely touched the model in order to make it persistent.