Test arch for handlers
Barely anything new other than the tornado test http client.

This is pretty much a dump of my git-stash. Let me get some work done now.

The plan is to finish the task api (and hopefully create a simple runner) and
leave the cosmetics for some day during the week.