Adding WTForms as a dep
Still a lot of unwanted boilerplate in the requirements.txt. I am just doing
some experiments with WTForms and how it integrates with the tornado and
sqlalchemy environment.

Actually I am struggling quite a lot with this. I thought it would be much
simpler. I guess I am just too old and grumpy.