e2e task api tests
Oh dear god. After a lot of struggling I finally got the async client tests to
work. Biggest gotcha of the week: I was adding tasks to the DB and not
flushing, thus my tests had side effects that drove me crazy.

Good news is that I've also serialized my objects in a nice json output. Plan
of the day is to write a not-so-dumb test to /get and move on to the worker

yay!