create/list task GUI.
+ Finished the task api tests.

+ New controller to trigger new task creation

+ Template hierarchy (with a lazy WTForm macro for Jinja2:
http://flask.pocoo.org/docs/0.10/patterns/wtforms/#forms-in-templates)

~ General controller refactoring

-

I think I got to a good architecture to scale the API/GUI with the business
logic. I ended up getting my modules a bit dirty in the dependencies lately but
it has come the time that I just wanna things working and keep refactoring. My
tests has been catching most of the bugs so far, which is great :-)

WTForms, WTForms-Alchemy, Tornado and Jinja2 are starting to play together...
Hopefully my job gets a bit faster in the future.

On the personal side: vacations until monday. ┻━┻︵ \(°□°)/ ︵ ┻━┻