Document modules that are fairly stable
Adds some documentation to the task controllers and model as I think these are
close to stable. At least their responsibilities shouldn't change in the near
future.

~

I'm too  busy working on the consumer/producer job to write a "dear diary"
message for this commit. But things are looking great :)