Base job runner in the works
Just checking in the first version of the cronjob runner. A producer/consumer
that takes advantage of the tornado.queue mechanism.

~

This is getting real. Most likely I'll have a working product after the next
coding session. And then I'll give it a break and work on docs and interface
cuteness before adding fancy matching and remaining features.