Runners in place. First integration test as well
+ cronjob script to run tasks.

~ worker refactor

+ brand new /api/run to run all tasks (filters coming next)

+ first integration test! yay! oops, I just realized that I'm not mocking the
servers. that explains 4s to run the tests :P

~

After a long hiatus working my ass off in my 9to5 I finally got some time to
fix the race condition that was hapenning on test_integration. I've gotta
confess I was trying brute force changing random one-liners in the worker
trying to organize the way the waits()  on tornado would deal with sqlalchemy
updating my models.

I decided to keep it simple and wrote a new route (which is going to be useful
in the future, as we'll have a web interface to run one or many tasks).

For now, I am calling it a day and should have a full e2e with field validation
and fake backends before I start getting blacklisted from places.

-
bigo