POST,PUT, DELETE and GET without filters are OK.
+ A few tweaks here and there to make the tests via cURL match the unittests.

~ I did a first full-flow test with real cURL requests and made these work as
well...

~

Hellah work that could have been avoided if I was in a good shape with my
python skills. I start to hate the magic that wtforms-sqlalchemy tries to make.
It is twice as hard to debug things with this. Hopefully it is just temporary
or it is just me :-)

New goal pre-thanks giving: implement a simple GET filter and write some docs
with what I have so far.

Docs should be part of the development rather than a step once everything is
done. I feel like now it is a good time to start investing on them.