─=≡Σ((( つ◕ل͜◕)つ We have docs!
+ First version of the docs.

~

Alright kids, we have our first version of a README that actually helps.

It reflects what is working and what are the plans for the future, as of
today. If things change tomorrow, I'll update it :)

It also has a nice disclaimer to avoid hurting people's feelings when I
decided to chop endpoints of change output formats