Switch to tornado.options and kills manage.py
~ Replaces the rudimentary config system that I had create with tornado.options
(and creates a /conf/ dir to host all the configurations != default)

~ Creates a test.py/run.py and retired manage.py. Updated some docs and my next
commit will probably be me fixing the travis job :)

~ Start using nosetests to manage my tests instead of plain unittest module.

~

Preparing the terrain for more configuration options. the in-line prompt
function.

Also added "coverage" as a dependency and running coverage after tests. Right
now it says 85% with mainly the controllers missing. Improves on the way.. but
not tonight.