Migrating GUI to client/
It is clear to me now that AngularJS will replace jinja2 as the project moves
forward. I love the idea of using ngResource as a client for /api. Pretty proud
of this setup right now.

This commit implement the "list.html" migration to the new architecture as well
as the angular-resource and angular-route deps.

!! This branch is still broken !!

~

There are probably some AngularJS rookie mistakes which I am gonna fix as I go.
I am pretty new to the framework but it's been a fun learning so far.