house cleaning and minimum_interval
A bit of house cleaning and small refactoring. I've also added
a "minimum_interval" to the configuration to prevent tasks to run in a too
short interval. Default is 15 minutes, chosen arbitrarily. Should revisit this
soon.

~

I've also got black listed on github api due to the amount of test requests.
yay.